alteração da estrutura banco

+++ trunk/Persistence/Entity/NotaFalta.cs
@@ -17,8 +17,8 @@ namespace Persistence.Entity
         public virtual int idModulo { get; set; }
         public virtual int idAluno { get; set; }
         public virtual int qtdFalta { get; set; }
-        public virtual Single nota1 { get; set; }
-        public virtual Single nota2 { get; set; }
+        public virtual decimal nota1 { get; set; }
+        public virtual decimal nota2 { get; set; }
         public virtual Aluno Aluno { get; set; }
         public virtual Modulo Modulo { get; set; }
         public virtual Turma Turma { get; set; }
